re-use regex loops
