+++ plugins/classify/src/byTitle/patterns.js
@@ -1,0 +1,4 @@
+module.exports = {
+  'CreativeWork/Film': [/ \([0-9]{4} film\)$/],
+  CreativeWork: [/ \((.*? )song\)$/]
+}
